// apps/sae/src/modules-features/admin/ModuleSystemSettings/StaffCategories/interfaces/IStaffInfoViewModel.ts
export interface IStaffInfoViewModel {
    id?: number;
    code?: string;
    lastName?: string;
    middleName?: string;
    firstName?: string;
    gender?: string;
    dateOfBirth?: Date;
    phoneNumber?: string;
    email?: string;
    position?: string;
    department?: string;
    isExternalStaff?: boolean;
} 